% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_estimate_fast.R
\name{Graph_Estimate_fast}
\alias{Graph_Estimate_fast}
\title{Fast Graph Estimation (single perturbation)}
\usage{
Graph_Estimate_fast(b_mat, se_mat, pair_info, random_start = 10, seed = 0)
}
\arguments{
\item{b_mat}{perturbed effect matrix}

\item{se_mat}{standard error matrix}

\item{pair_info}{list from \code{make_pair_info}}

\item{random_start}{passed to mr_cML_O}

\item{seed}{random seed}
}
\value{
list: obs_graph, obs_graph_se, obs_graph_pval
}
\description{
Fast Graph Estimation (single perturbation)
}
