% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_perturb_fast.R
\name{Graph_Perturb_fast}
\alias{Graph_Perturb_fast}
\title{Parallel graph perturbation (fast)}
\usage{
Graph_Perturb_fast(
  n_vec,
  rho_mat,
  screen_res,
  num_pert = 100,
  random_start = 10,
  seed = 0,
  trait_vec = NULL
)
}
\arguments{
\item{n_vec}{See \code{MR2G()}.}

\item{rho_mat}{See \code{MR2G()}.}

\item{screen_res}{A list that must contain \code{b_mat}, \code{se_mat},
\code{IJ_snp_list}, and optionally \code{DP_mat_list}.}

\item{num_pert}{Number of perturbations.}

\item{random_start}{See \code{MR2G()}.}

\item{seed}{Random seed.}

\item{trait_vec}{Optional trait names.}
}
\value{
List containing lists of \code{obs_graph}, \code{obs_graph_se}, \code{obs_graph_pval}, and \code{trait_vec}.
}
\description{
Parallel graph perturbation (fast)
}
