% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IV_list}
\alias{IV_list}
\title{Example IV list}
\format{
A list with elements:
\describe{
\item{trait1}{Character vector of SNPs}
\item{trait2}{Character vector of SNPs}
}
}
\usage{
IV_list
}
\description{
A list of instrumental variables for testing.
}
\examples{
data("IV_list")
}
\keyword{datasets}
